Rename result list page to snake case
