+++ lib/views/result_list_page.dart
